% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/example1.R
\name{example1}
\alias{example1}
\title{Calling a shiny app with inputs from the user's session}
\usage{
example1(Input1, Input2, Input3)
}
\arguments{
\item{Input1}{the unquoted name of an object from the session}

\item{Input2}{(optional) the unquoted name of an object from the session}

\item{Input3}{(optional) the unquoted name of an object from the session}
}
\description{
Calling a shiny app with inputs from the user's session
}

