% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simple_tab_panels.R
\name{example_panels}
\alias{example_panels}
\alias{simple_tab_panel_1}
\alias{simple_tab_panel_2}
\title{generate a very simple UI with a plot, a slider, and a selector}
\usage{
simple_tab_panel_1(title = "I'm the title", uniqueID)

simple_tab_panel_2(title = "number 2", uniqueID)
}
\arguments{
\item{uniqueID}{a character string serving as a unique ID for the
widgets in this panel.}
}
\description{
generate a very simple UI with a plot, a slider, and a selector
}

